fix: resolve 20 code review issues from deep codebase audit
CRITICAL:
- MainWindow.Dispose: CTS disposal leak, window handle race guard
- PathUtils.SanitizeFileName: trailing-period removal loop broken (unconditional break)
- args.Replace: use quoted-path matching to prevent substring corruption
- ReportException: non-blocking for dispatcher handler, sync only for AppDomain
- BugReportApiSink: 12s safety timer prevents indefinite throttle flag hang

HIGH:
- LoadFilesForConversion/Verification/Extraction: catch OCE on cancellation
- ScreenshotService: nested try/finally prevents GDI handle leaks
- ConvertToChdAsync: recursion depth guard on createdvd retry
- CsoFile/PbpFile/PbpDiscInfo: BitConverter -> BinaryPrimitives for endianness
- CueBinWriter: retry logic for File.Copy calls
- SafeFireAndForget: removed broken early-return on IsCompleted

MEDIUM:
- ResolvedInput: null guard before null-forgiving operator
- CueSheetWriter.SubtractLeadin: long arithmetic prevents overflow

LOW:
- ArchiveService: removed empty Dispose (no unmanaged resources)
- FakeHttpMessageHandler: removed confusing dummy handler
- ArchiveServiceTests: removed DisposeDoesNotThrow test
- App.xaml.cs: string comparison MA0006 warning

+++ PBPSharp/CueSheetWriter.cs
@@ -46,16 +46,16 @@ public static class CueSheetWriter
 
     private static (int Minutes, int Seconds, int Frames) SubtractLeadin(int minutes, int seconds, int frames, int leadinFrames)
     {
-        var totalFrames = (minutes * 60 * 75) + (seconds * 75) + frames - leadinFrames;
+        var totalFrames = (long)minutes * 60 * 75 + (long)seconds * 75 + frames - leadinFrames;
         if (totalFrames < 0)
         {
             totalFrames = 0;
         }
 
-        var m = totalFrames / (60 * 75);
+        var m = (int)(totalFrames / (60 * 75));
         var remainder = totalFrames % (60 * 75);
-        var s = remainder / 75;
-        var f = remainder % 75;
+        var s = (int)(remainder / 75);
+        var f = (int)(remainder % 75);
 
         return (m, s, f);
     }
